moved LightClient mods to files
